chatbot_api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
sentence-transformers
qdrant-client
psycopg2-binary
beautifulsoup4
markdown
pydantic-settings